Safari-609.1.11.1/LayoutTests/http/tests/globalhistory/history-delegate-pushstate-expected.txt

WebView navigated to url "http://127.0.0.1:8000/globalhistory/history-delegate-pushstate.html" with title "" with HTTP equivalent method "GET".  The navigation was successful and was not a client redirect.
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
WebView navigated to url "http://127.0.0.1:8000/globalhistory/history-delegate-pushstate.html?1" with title "" with HTTP equivalent method "GET".  The navigation was successful and was not a client redirect.
WebView navigated to url "http://127.0.0.1:8000/globalhistory/history-delegate-pushstate.html?2" with title "" with HTTP equivalent method "GET".  The navigation was successful and was not a client redirect.
WebView navigated to url "http://127.0.0.1:8000/globalhistory/history-delegate-pushstate.html?3" with title "" with HTTP equivalent method "GET".  The navigation was successful and was not a client redirect.
WebView navigated to url "http://127.0.0.1:8000/globalhistory/history-delegate-pushstate.html?4" with title "" with HTTP equivalent method "GET".  The navigation was successful and was not a client redirect.
WebView navigated to url "http://127.0.0.1:8000/globalhistory/history-delegate-pushstate.html?5" with title "" with HTTP equivalent method "GET".  The navigation was successful and was not a client redirect.
This tests functionality of the history delegate related to history.pushState.
